+++ apiserver/paasng/tests/paas_wl/bk_app/cnative/specs/test_resource.py
@@ -67,7 +67,7 @@ class TestMresConditionDetector:
                     ],
                     MResPhaseType.AppFailed,
                 ),
-                DeployStatus.READY,
+                DeployStatus.ERROR,
             ),
         ],
     )
