+++ operator/pkg/controllers/reconcilers/finalizer.go
@@ -55,7 +55,7 @@ func (r *BkappFinalizer) Reconcile(ctx context.Context, bkapp *paasv1alpha2.BkAp
 	// our finalizer is present, so lets handle any external dependency
 	finished, err := r.hooksFinished(ctx, bkapp)
 	if err != nil {
-		return r.Result.withError(errors.Wrap(err, "failed to check hook status"))
+		return r.Result.WithError(errors.Wrap(err, "failed to check hook status"))
 	}
 	if !finished {
 		apimeta.SetStatusCondition(&bkapp.Status.Conditions, metav1.Condition{
@@ -70,13 +70,13 @@ func (r *BkappFinalizer) Reconcile(ctx context.Context, bkapp *paasv1alpha2.BkAp
 	if err = r.deleteResources(ctx, bkapp); err != nil {
 		// if fail to delete the external dependency here, return with error
 		// so that it can be retried
-		return r.Result.withError(errors.Wrap(err, "failed to delete external resources"))
+		return r.Result.WithError(errors.Wrap(err, "failed to delete external resources"))
 	}
 
 	// remove our finalizer from the finalizers list and update it.
 	controllerutil.RemoveFinalizer(bkapp, paasv1alpha2.BkAppFinalizerName)
 	if err = r.Client.Update(ctx, bkapp); err != nil {
-		return r.Result.withError(errors.Wrap(err, "failed to remove finalizer for app"))
+		return r.Result.WithError(errors.Wrap(err, "failed to remove finalizer for app"))
 	}
 	return r.Result.End()
 }
