+++ operator/pkg/controllers/reconcilers/services.go
@@ -48,7 +48,7 @@ type ServiceReconciler struct {
 func (r *ServiceReconciler) Reconcile(ctx context.Context, bkapp *paasv1alpha2.BkApp) Result {
 	current, err := r.listCurrentServices(ctx, bkapp)
 	if err != nil {
-		return r.Result.withError(err)
+		return r.Result.WithError(err)
 	}
 
 	expected := r.getWantedService(bkapp)
@@ -57,13 +57,13 @@ func (r *ServiceReconciler) Reconcile(ctx context.Context, bkapp *paasv1alpha2.B
 	if len(outdated) != 0 {
 		for _, svc := range outdated {
 			if err = r.Client.Delete(ctx, svc); err != nil {
-				return r.Result.withError(err)
+				return r.Result.WithError(err)
 			}
 		}
 	}
 	for _, svc := range expected {
 		if err = r.applyService(ctx, svc); err != nil {
-			return r.Result.withError(err)
+			return r.Result.WithError(err)
 		}
 	}
 	return r.Result
