--- operator/pkg/controllers/reconcilers/status.go
@@ -1,49 +1,0 @@
-/*
- * TencentBlueKing is pleased to support the open source community by making
- * 蓝鲸智云 - PaaS 平台 (BlueKing - PaaS System) available.
- * Copyright (C) 2017 THL A29 Limited, a Tencent company. All rights reserved.
- * Licensed under the MIT License (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- *	http://opensource.org/licenses/MIT
- *
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * We undertake not to change the open source license (MIT license) applicable
- * to the current version of the project delivered to anyone in the future.
- */
-
-package reconcilers
-
-import "context"
-
-import (
-	"fmt"
-
-	paasv1alpha2 "bk.tencent.com/paas-app-operator/api/v1alpha2"
-
-	"github.com/pkg/errors"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-)
-
-// UpdateStatus update the status of BkApp
-func UpdateStatus(ctx context.Context, client client.Client, bkapp *paasv1alpha2.BkApp, ret Result) Result {
-	if ret.err == nil {
-		// 更新 ObservedGeneration, 表示成功完成了一次新版本的调和流程
-		bkapp.Status.ObservedGeneration = bkapp.Generation
-	}
-
-	if err := client.Status().Update(ctx, bkapp); err != nil {
-		syncErr := errors.Wrap(err, "failed to update bkapp status")
-		if ret.err == nil {
-			return ret.withError(syncErr)
-		} else {
-			return ret.withError(fmt.Errorf("%s; %s", syncErr, ret.err))
-		}
-	}
-
-	return ret
-}
